refactor(api modal): fix
