+++ src/components/burger-contructor/burger-constructor.tsx
@@ -6,8 +6,8 @@ import {
 	Button,
 	CurrencyIcon,
 } from '@ya.praktikum/react-developer-burger-ui-components';
-import { Modal } from '../modal-components/modal/modal';
 import { OrderDetails } from './order-details/order-details';
+import { Modal } from '../modal/modal';
 
 type TBurgerConstructorProps = {
 	ingredients: TIngredient[];
