+++ src/utils/api.ts
@@ -1,14 +1,14 @@
 import { TIngredient } from './types';
 
+const BURGER_API_URL = 'https://norma.nomoreparties.space/api';
+
 type TIngredientsResponse = {
 	success: boolean;
 	data: TIngredient[];
 };
 
 export const getIngredients = async (): Promise<TIngredientsResponse> => {
-	const url = 'https://norma.nomoreparties.space/api/ingredients';
-
-	const res = await fetch(url);
+	const res = await fetch(`${BURGER_API_URL}/ingredients`);
 	const data = await res.json();
 	return data;
 };
